+++ src/omplapp/apps/QuadrotorPlanning.cpp
@@ -53,7 +53,7 @@ ompl::base::ScopedState<> ompl::app::QuadrotorPlanning::getFullStateFromGeometri
 void ompl::app::QuadrotorPlanning::propagate(const base::State *from, const control::Control *ctrl,
     const double duration, base::State *result)
 {
-    int i, j, nsteps = ceil(duration/timeStep_);
+    int i, j, nsteps = static_cast<int>(ceil(duration/timeStep_));
     double dt = duration/(double)nsteps;
     base::State *dstate = getStateSpace()->allocState();
     base::CompoundStateSpace::StateType&      s = *result->as<base::CompoundStateSpace::StateType>();
